+++ spec/factories/mailchimp_member_factories.rb
@@ -4,13 +4,11 @@ require 'ffcrm_mailchimp/member'
 FactoryGirl.define do
 
   factory :mailchimp_member, class: FfcrmMailchimp::Member do
-    email { Faker::Internet.email }
-    first_name { Faker::Name.first_name }
-    last_name  { Faker::Name.last_name }
-    list_id { '12345' }
-    subscribed_groups { {'Group One' => "Option 1, Option 2", "Group Two" => "Option 3" } }
-    last_changed { DateTime.now }
-    phone { Faker::PhoneNumber.phone_number }
+    email_address   { Faker::Internet.email }
+    status          { "subscribed" }
+    list_id         { '12345' }
+    merge_fields    { {"FIRST_NAME" => Faker::Name.first_name, "LAST_NAME" => Faker::Name.last_name} }
+    interests       { {"70b7107c8a" => true, "7c1719c788" => false, "8d856390f6" => true} }
     initialize_with { attributes }
   end