+++ spec/lib/ffcrm_mailchimp/api_spec.rb
@@ -1,0 +1,23 @@
+require 'spec_helper'
+
+describe FfcrmMailchimp::Api do
+
+  describe "all_lists" do
+  end
+
+  describe "subscribe" do
+  end
+
+  describe "unsubscribe" do
+  end
+
+  describe "interest_groupings" do
+  end
+
+  describe "lookup_member_on_list" do
+  end
+
+  describe "clear_cache" do
+  end
+
+end